% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regnskapsanalyse.r
\name{analyser_regnskap}
\alias{analyser_regnskap}
\title{Analyser regnskapstall fra institusjon}
\usage{
analyser_regnskap(
  kilde_fil,
  output_fil,
  søkenavn,
  type = c("hogskole", "fagskole", "studentsamskipnad"),
  søke_ark = "Nøkkeltall"
)
}
\arguments{
\item{kilde_fil}{Filnavn for kildefilen}

\item{output_fil}{Filnavn for målfilen}

\item{søkenavn}{Navn på institusjonen som skal søkes etter i målarket}

\item{type}{Type institusjon: "hogskole", "fagskole" eller "studentsamskipnad"}

\item{søke_ark}{Navn på arket i målfilen der verdiene skal plasseres}
}
\value{
Usynlig retur, skriver data til målfil
}
\description{
Analyser regnskapstall fra institusjon
}
