% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fagskole.r
\name{fagskole_analyse}
\alias{fagskole_analyse}
\title{Analyser regnskapstall fra fagskole}
\usage{
fagskole_analyse(kilde_fil, output_fil, søkenavn, søke_ark = "Nøkkeltall")
}
\arguments{
\item{kilde_fil}{Filnavn for kildefilen}

\item{output_fil}{Filnavn for målfilen}

\item{søkenavn}{Navn på fagskolen som skal søkes etter i målarket}

\item{søke_ark}{Navn på arket i målfilen der verdiene skal plasseres}
}
\value{
Usynlig retur, skriver data til målfil
}
\description{
Analyser regnskapstall fra fagskole
}
