% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{finn_institusjon_rad}
\alias{finn_institusjon_rad}
\title{Finn eller opprett rad for institusjon i målarket}
\usage{
finn_institusjon_rad(mål_data, søkenavn)
}
\arguments{
\item{mål_data}{Datarammen for målarket}

\item{søkenavn}{Navnet på institusjonen}
}
\value{
Radnummer
}
\description{
Finn eller opprett rad for institusjon i målarket
}
