% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{finn_omlopsmidler}
\alias{finn_omlopsmidler}
\title{Finn omløpsmidler med alternative metoder}
\usage{
finn_omlopsmidler(eiendeler_data, omlopsmidler_komponenter)
}
\arguments{
\item{eiendeler_data}{Dataramme med eiendelsdata}

\item{omlopsmidler_komponenter}{Vektor med komponenter for omløpsmidler}
}
\value{
Liste med funnet-flagg og eventuell verdi
}
\description{
Finn omløpsmidler med alternative metoder
}
