% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{finn_verdi}
\alias{finn_verdi}
\title{Søk etter verdi for nøkkelord i dataramme}
\usage{
finn_verdi(data, nøkkelord)
}
\arguments{
\item{data}{Datarammen som skal søkes i}

\item{nøkkelord}{Nøkkelordet som skal finnes}
}
\value{
Liste med funnet-flagg og eventuell verdi
}
\description{
Søk etter verdi for nøkkelord i dataramme
}
