% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{konverter_til_numerisk}
\alias{konverter_til_numerisk}
\title{Konverter tekstverdi til numerisk}
\usage{
konverter_til_numerisk(verdi)
}
\arguments{
\item{verdi}{Verdien som skal konverteres}
}
\value{
Numerisk verdi
}
\description{
Konverter tekstverdi til numerisk
}
