% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studentsamskipnad.r
\name{studentsamskipnad_analyse}
\alias{studentsamskipnad_analyse}
\title{Analyser regnskapstall fra studentsamskipnad}
\usage{
studentsamskipnad_analyse(
  kilde_fil,
  output_fil,
  søkenavn,
  søke_ark = "Nøkkeltall"
)
}
\arguments{
\item{kilde_fil}{Filnavn for kildefilen}

\item{output_fil}{Filnavn for målfilen}

\item{søkenavn}{Navn på studentsamskipnaden som skal søkes etter i målarket}

\item{søke_ark}{Navn på arket i målfilen der verdiene skal plasseres}
}
\value{
Usynlig retur, skriver data til målfil
}
\description{
Analyser regnskapstall fra studentsamskipnad
}
